Implement basic gdb server functionality.
This commit introduces basic gdbserver support to the zmu ARM simulator,
enabling remote debugging with a gdb client.

The following features are implemented:

* Breakpoints: Users can set, clear, and manage breakpoints during program execution.
* Continue: Execution can be resumed from the current breakpoint or paused state.
* Step Instruction: Users can step through program instructions for detailed debugging.

This functionality significantly enhances the debugging capabilities of zmu,
making it more versatile for developers.

To start the gdbserver just call zmu with the --gdb flag:

$ zmu.exe run --gdb binary.elf

A gdb server will be open on localhost port 9001

Signed-off-by: Diego Asanza <f.asanza@de.abb.com>