Fix execution priority.
This commit implements the execution priority as defined in the
armv7-m architecture reference manual.

There was a bug where the group priority for reset, nmi and hardfault
was not being ignored as specified in the ARMV7-M Arch Reference Manual.
Version E. Page B1-157. Here it is defined that the group priority
for Hardfault, NMI and Reset is -1, -2 and -3 respectively, regardles
of the value of PRIGROUP.

Signed-off-by: Diego Asanza <f.asanza@de.abb.com>